refactor: renamed config to AppConfig
