+++ sites/geohub/src/components/DataView.svelte
@@ -5,13 +5,14 @@
   import { map, indicatorProgress } from '$stores'
   import TextFilter from '$components/data-view/TextFilter.svelte'
   import Notification from '$components/controls/Notification.svelte'
-  import { DataCategories, STAC_MINIMUM_ZOOM, tagSearchKeys } from '$lib/constants'
+  import { STAC_MINIMUM_ZOOM } from '$lib/constants'
   import DataCategoryCardList from '$components/data-view/DataCategoryCardList.svelte'
   import { Breadcrumbs, Loader, type Breadcrumb } from '@undp-data/svelte-undp-design'
   import type { Tag } from '$lib/types/Tag'
   import SelectedTags from './data-view/SelectedTags.svelte'
   import { goto } from '$app/navigation'
   import { getSelectedTagsFromUrl } from '$lib/helper'
+  import { DataCategories, TagSearchKeys } from '$lib/AppConfig'
 
   const session = $page.data.session
   const dataCategories: Breadcrumb[] = session
@@ -130,7 +131,7 @@
   const handleTagChanged = async (e) => {
     selectedTags = e.detail.tags
     const apiUrl = $page.url
-    tagSearchKeys.forEach((key) => {
+    TagSearchKeys.forEach((key) => {
       apiUrl.searchParams.delete(key.key)
     })
     selectedTags?.forEach((t) => {
@@ -248,7 +249,7 @@
   }
 
   const clearSelectedTags = async (url: URL) => {
-    tagSearchKeys.forEach((key) => {
+    TagSearchKeys.forEach((key) => {
       url.searchParams.delete(key.key)
     })
     selectedTags = []
