+++ sites/geohub/src/components/Map.svelte
@@ -14,11 +14,11 @@
   import MapQueryInfoControl from '$components/MapQueryInfoControl.svelte'
   import StyleSwicher from '@undp-data/style-switcher'
   import CurrentLocation from '@undp-data/current-location'
-  import { MAP_ATTRIBUTION, styles } from '$lib/constants'
   import { loadImageToDataUrl, fetchUrl, clipSprite } from '$lib/helper'
   import type { Sprite } from '$lib/types'
   import { spriteImageList } from '$stores'
   import LayerVisibilitySwitcher from './LayerVisibilitySwitcher.svelte'
+  import { attribution, MapStyles } from '$lib/AppConfig'
 
   let container: HTMLDivElement
   export let map: Map
@@ -35,14 +35,14 @@
     return new Promise<void>((resolve) => {
       map = new Map({
         container,
-        style: styles[0].uri,
+        style: MapStyles[0].uri,
         center: [0, 0],
         zoom: 3,
         hash: true,
         attributionControl: false,
       })
 
-      map.addControl(new AttributionControl({ compact: true, customAttribution: MAP_ATTRIBUTION }), 'bottom-right')
+      map.addControl(new AttributionControl({ compact: true, customAttribution: attribution }), 'bottom-right')
       map.addControl(
         new NavigationControl({
           visualizePitch: true,
@@ -132,7 +132,7 @@
   <MapQueryInfoControl bind:map />
   <StyleSwicher
     bind:map
-    {styles}
+    styles={MapStyles}
     position="bottom-left" />
   <LayerVisibilitySwitcher
     bind:map
