+++ sites/geohub/src/components/maps/MapStyleCard.svelte
@@ -7,7 +7,7 @@
   import { clickOutside } from 'svelte-use-click-outside'
   import { Accordion, Button, CtaLink, Loader } from '@undp-data/svelte-undp-design'
   import type { DashboardMapStyle } from '$lib/types'
-  import { AccessLevel } from '$lib/constants'
+  import { AccessLevel } from '$lib/AppConfig'
   import { sleep } from '$lib/helper'
 
   const dispatch = createEventDispatcher()
