+++ sites/geohub/src/lib/AppConfig/AccessLevel.ts
@@ -1,0 +1,5 @@
+export enum AccessLevel {
+  PRIVATE = 1,
+  ORGANIZATION = 2,
+  PUBLIC = 3,
+}
