+++ sites/geohub/src/lib/AppConfig/DatasetSearchQueryParams.ts
@@ -1,0 +1,12 @@
+export const DatasetSearchQueryParams = [
+  'query',
+  'offset',
+  'limit',
+  'storage_id',
+  'bbox',
+  'sortby',
+  'operator',
+  'staronly',
+  'queryoperator',
+  'mydata',
+]
