+++ sites/geohub/src/lib/AppConfig/DatasetSortingColumns.ts
@@ -1,0 +1,24 @@
+import type { Radio } from '@undp-data/svelte-undp-design'
+
+export const DatasetSortingColumns: Radio[] = [
+  {
+    value: 'name,asc',
+    label: 'A to Z',
+  },
+  {
+    value: 'name,desc',
+    label: 'Z to A',
+  },
+  {
+    value: 'no_stars,desc',
+    label: 'Most favourite',
+  },
+  {
+    value: 'updatedat,desc',
+    label: 'Most recent',
+  },
+  {
+    value: 'updatedat,asc',
+    label: 'Less recent',
+  },
+]
