+++ sites/geohub/src/lib/AppConfig/FooterItems.ts
@@ -1,0 +1,38 @@
+export const FooterItems: { [key: string]: { title: string; url: string }[] } = {
+  GeoHub: [
+    {
+      title: 'GeoHub',
+      url: 'https://geohub.data.undp.org',
+    },
+    {
+      title: 'Shared Maps',
+      url: 'https://geohub.data.undp.org/maps',
+    },
+    {
+      title: 'Dashboards',
+      url: 'https://geohub.data.undp.org/dashboards',
+    },
+    {
+      title: 'Electricity Dashboard',
+      url: 'https://geohub.data.undp.org/dashboard/electricity',
+    },
+    {
+      title: 'User Guide',
+      url: 'https://docs.undpgeohub.org',
+    },
+  ],
+  'For Developers': [
+    {
+      title: 'GeoHub Dev',
+      url: 'https://dev.undpgeohub.org',
+    },
+    {
+      title: 'GeoHub API documentation',
+      url: '/api',
+    },
+    {
+      title: 'Svelte UNDP design sytem',
+      url: 'https://svelte-undp-design.undpgeohub.org',
+    },
+  ],
+}
