+++ sites/geohub/src/lib/AppConfig/LimitOptions.ts
@@ -1,0 +1,1 @@
+export const LimitOptions = [5, 10, 25, 50, 100]
