+++ sites/geohub/src/lib/AppConfig/MapSortingColumns.ts
@@ -1,0 +1,18 @@
+export const MapSortingColumns = [
+  {
+    value: 'updatedat,desc',
+    label: 'Most recent',
+  },
+  {
+    value: 'updatedat,asc',
+    label: 'less recent',
+  },
+  {
+    value: 'name,asc',
+    label: 'A to Z',
+  },
+  {
+    value: 'name,desc',
+    label: 'Z to A',
+  },
+]
