+++ sites/geohub/src/lib/AppConfig/MapStyles.ts
@@ -1,0 +1,12 @@
+import type { StyleDefinition } from '@undp-data/style-switcher'
+
+export const MapStyles: StyleDefinition[] = [
+  {
+    title: 'Carto',
+    uri: 'https://undp-data.github.io/style/style.json',
+  },
+  {
+    title: 'Bing Aerial',
+    uri: 'https://undp-data.github.io/style/aerialstyle.json',
+  },
+]
