+++ sites/geohub/src/lib/AppConfig/Permission.ts
@@ -1,0 +1,5 @@
+export enum Permission {
+  READ = 1,
+  WRITE = 2,
+  OWNER = 3,
+}
