+++ sites/geohub/src/lib/AppConfig/TagInputValues.ts
@@ -1,0 +1,30 @@
+export const TagInputValues = [
+  {
+    key: 'extent',
+    label: 'Extent',
+  },
+  {
+    key: 'granularity',
+    label: 'Admin Level',
+  },
+  {
+    key: 'resolution',
+    label: 'Resolution',
+  },
+  {
+    key: 'year_value',
+    label: 'Year',
+  },
+  {
+    key: 'sdg_target',
+    label: 'SDG Topic',
+  },
+  {
+    key: 'theme',
+    label: 'Theme',
+  },
+  {
+    key: 'unit',
+    label: 'Unit',
+  },
+]
