+++ sites/geohub/src/lib/AppConfig/TagSearchKeys.ts
@@ -1,0 +1,54 @@
+export const TagSearchKeys = [
+  {
+    key: 'extent',
+    label: 'Extent',
+  },
+  {
+    key: 'continent',
+    label: 'Continent',
+  },
+  {
+    key: 'region',
+    label: 'Region',
+  },
+  {
+    key: 'country',
+    label: 'Country',
+  },
+  {
+    key: 'granularity',
+    label: 'Admin Level',
+  },
+  {
+    key: 'resolution',
+    label: 'Resolution',
+  },
+  {
+    key: 'year_value',
+    label: 'Year',
+  },
+  {
+    key: 'keyword',
+    label: 'Keyword',
+  },
+  {
+    key: 'sdg_target',
+    label: 'SDG Topic',
+  },
+  {
+    key: 'theme',
+    label: 'Theme',
+  },
+  {
+    key: 'schema',
+    label: 'Schema',
+  },
+  {
+    key: 'provider',
+    label: 'Data Provider',
+  },
+  {
+    key: 'type',
+    label: 'GIS Server',
+  },
+]
