+++ sites/geohub/src/lib/AppConfig/attribution.ts
@@ -1,0 +1,2 @@
+export const attribution =
+  '<a target="_top" rel="noopener" href="http://undp.org">United Nations Development Programme (UNDP)</a>'
