+++ sites/geohub/src/lib/constants.ts
@@ -1,7 +1,3 @@
-import type { StyleDefinition } from '@undp-data/style-switcher'
-import type { Breadcrumb, Radio } from '@undp-data/svelte-undp-design'
-import type { AcceptedExtension } from './types'
-
 export const COLOR_CLASS_COUNT = 5
 export const COLOR_CLASS_COUNT_MAXIMUM = 25
 export const COLOR_CLASS_COUNT_MINIMUM = 2
@@ -12,9 +8,6 @@ export const SEARCH_PAGINATION_LIMIT = 25
 export const TOKEN_EXPIRY_PERIOD_MSEC = 86400000 * 365 // for 1 year
 export const STAC_MINIMUM_ZOOM = 5
 export const UNIQUE_VALUE_THRESHOLD = 25
-
-export const MAP_ATTRIBUTION =
-  '<a target="_top" rel="noopener" href="http://undp.org">United Nations Development Programme (UNDP)</a>'
 
 export enum TabNames {
   ANALYZE = 'Analyze',
@@ -76,25 +69,7 @@ export enum LayerTypes {
   HEATMAP = 'heatmap',
 }
 
-export enum AccessLevel {
-  PRIVATE = 1,
-  ORGANIZATION = 2,
-  PUBLIC = 3,
-}
-
-export const LimitOptions = [5, 10, 25, 50, 100]
 export const DEFAULT_LIMIT = 10
-
-export const styles: StyleDefinition[] = [
-  {
-    title: 'Carto',
-    uri: 'https://undp-data.github.io/style/style.json',
-  },
-  {
-    title: 'Bing Aerial',
-    uri: 'https://undp-data.github.io/style/aerialstyle.json',
-  },
-]
 
 export const vectorFilterOperations = [
   { value: '==', label: 'Equals', text: 'is equal to', disabled: false, symbol: '=' },
@@ -120,258 +95,4 @@ export const rasterArithmeticOperators = [
   { value: '/', label: 'Divide', text: 'division', disabled: false, symbol: '/' },
   { value: '%', label: 'Modulo', text: 'remainder of division', disabled: false, symbol: '%' },
   { value: '**', label: 'Power', text: 'raise to power', disabled: false, symbol: '^' },
-]
-
-export const DataCategories: Breadcrumb[] = [
-  {
-    name: 'SDG',
-    icon: 'assets/sdgs/SDG Wheel_WEB.png',
-    url: '/api/tags?key=sdg_goal',
-  },
-  {
-    name: 'UNDP',
-    icon: 'assets/undp-images/undp-logo.png',
-    url: '/api/datasets?provider=undp',
-  },
-  {
-    name: 'Microsoft Planetary',
-    icon: 'assets/microsoft.png',
-    url: '/api/datasets?stac=microsoft-pc',
-  },
-  {
-    name: 'Dynamic vector data',
-    icon: 'assets/postgresql.png',
-    url: '/api/datasets?type=pgtileserv',
-  },
-  {
-    name: 'Favourite',
-    icon: 'assets/star.png',
-    url: '/api/datasets?staronly=true',
-  },
-  {
-    name: 'My data',
-    icon: 'fa-solid fa-circle-user',
-    url: '/api/datasets?mydata=true',
-  },
-]
-
-export const SortingColumns: Radio[] = [
-  {
-    value: 'name,asc',
-    label: 'A to Z',
-  },
-  {
-    value: 'name,desc',
-    label: 'Z to A',
-  },
-  {
-    value: 'no_stars,desc',
-    label: 'Most favourite',
-  },
-  {
-    value: 'updatedat,desc',
-    label: 'Most recent',
-  },
-  {
-    value: 'updatedat,asc',
-    label: 'Less recent',
-  },
-]
-
-export const DatasetSearchQueryParams = [
-  'query',
-  'offset',
-  'limit',
-  'storage_id',
-  'bbox',
-  'sortby',
-  'operator',
-  'staronly',
-  'queryoperator',
-  'mydata',
-]
-
-export const tagSearchKeys = [
-  {
-    key: 'extent',
-    label: 'Extent',
-  },
-  {
-    key: 'continent',
-    label: 'Continent',
-  },
-  {
-    key: 'region',
-    label: 'Region',
-  },
-  {
-    key: 'country',
-    label: 'Country',
-  },
-  {
-    key: 'granularity',
-    label: 'Admin Level',
-  },
-  {
-    key: 'resolution',
-    label: 'Resolution',
-  },
-  {
-    key: 'year_value',
-    label: 'Year',
-  },
-  {
-    key: 'keyword',
-    label: 'Keyword',
-  },
-  {
-    key: 'sdg_target',
-    label: 'SDG Topic',
-  },
-  {
-    key: 'theme',
-    label: 'Theme',
-  },
-  {
-    key: 'schema',
-    label: 'Schema',
-  },
-  {
-    key: 'provider',
-    label: 'Data Provider',
-  },
-  {
-    key: 'type',
-    label: 'GIS Server',
-  },
-]
-
-export const tagInputValues = [
-  {
-    key: 'extent',
-    label: 'Extent',
-  },
-  {
-    key: 'granularity',
-    label: 'Admin Level',
-  },
-  {
-    key: 'resolution',
-    label: 'Resolution',
-  },
-  {
-    key: 'year_value',
-    label: 'Year',
-  },
-  {
-    key: 'sdg_target',
-    label: 'SDG Topic',
-  },
-  {
-    key: 'theme',
-    label: 'Theme',
-  },
-  {
-    key: 'unit',
-    label: 'Unit',
-  },
-]
-
-export const footerItems: { [key: string]: { title: string; url: string }[] } = {
-  GeoHub: [
-    {
-      title: 'GeoHub',
-      url: 'https://geohub.data.undp.org',
-    },
-    {
-      title: 'Shared Maps',
-      url: 'https://geohub.data.undp.org/maps',
-    },
-    {
-      title: 'Dashboards',
-      url: 'https://geohub.data.undp.org/dashboards',
-    },
-    {
-      title: 'Electricity Dashboard',
-      url: 'https://geohub.data.undp.org/dashboard/electricity',
-    },
-    {
-      title: 'User Guide',
-      url: 'https://docs.undpgeohub.org',
-    },
-  ],
-  'For Developers': [
-    {
-      title: 'GeoHub Dev',
-      url: 'https://dev.undpgeohub.org',
-    },
-    {
-      title: 'GeoHub API documentation',
-      url: '/api',
-    },
-    {
-      title: 'Svelte UNDP design sytem',
-      url: 'https://svelte-undp-design.undpgeohub.org',
-    },
-  ],
-}
-
-export const MapOrderByOptions = [
-  {
-    value: 'updatedat,desc',
-    label: 'Most recent',
-  },
-  {
-    value: 'updatedat,asc',
-    label: 'less recent',
-  },
-  {
-    value: 'name,asc',
-    label: 'A to Z',
-  },
-  {
-    value: 'name,desc',
-    label: 'Z to A',
-  },
-]
-
-export enum Permission {
-  READ = 1,
-  WRITE = 2,
-  OWNER = 3,
-}
-
-export const AccepedExtensions: AcceptedExtension[] = [
-  {
-    name: 'GeoTIFF',
-    extensions: ['tif', 'tiff'],
-    href: 'https://gdal.org/drivers/raster/gtiff.html#raster-gtiff',
-  },
-  {
-    name: 'NetCDF',
-    extensions: ['nc'],
-    href: 'https://gdal.org/drivers/raster/netcdf.html#raster-netcdf',
-  },
-  {
-    name: 'Arc/Info ASCII Grid File',
-    extensions: ['aig', 'asc', 'sgr', 'grd'],
-    href: 'https://gdal.org/drivers/raster/aaigrid.html',
-  },
-  { name: 'Erdas Imagine', extensions: ['raw', 'bl'], href: 'https://gdal.org/drivers/raster/eir.html' },
-  { name: 'ESRI Shapefile (zipped)', extensions: ['zip'], href: 'https://gdal.org/drivers/vector/shapefile.html' },
-  { name: 'GeoJSON', extensions: ['geojson'], href: 'https://gdal.org/drivers/vector/geojson.html' },
-  { name: 'PMTILES', extensions: ['pmtiles'], href: 'https://protomaps.com/docs/pmtiles' },
-  { name: 'MBTILES', extensions: ['mbtiles'], href: 'https://github.com/mapbox/mbtiles-spec' },
-  {
-    name: 'ESRI File Geodatabase (OpenFileGDB)',
-    extensions: ['gdb'],
-    href: 'https://gdal.org/drivers/vector/openfilegdb.html#esri-file-geodatabase-openfilegdb',
-  },
-  {
-    name: 'ESRI File Geodatabase (FileGDB)',
-    extensions: ['gdb'],
-    href: 'https://gdal.org/drivers/vector/filegdb.html',
-  },
-  { name: 'ESRI Personal GeoDatabase', extensions: ['mdb'], href: 'https://gdal.org/drivers/vector/pgeo.html' },
-  { name: 'GeoPackage', extensions: ['gpkg'], href: 'https://gdal.org/drivers/vector/gpkg.html' },
 ]
