+++ sites/geohub/src/lib/helper/createAttributionFromTags.ts
@@ -1,4 +1,4 @@
-import { MAP_ATTRIBUTION } from '$lib/constants'
+import { attribution } from '$lib/AppConfig'
 
 export const createAttributionFromTags = (tags: [{ key: string; value: string }]) => {
   const providers = tags?.filter((t) => t.key === 'provider')
@@ -7,6 +7,6 @@ export const createAttributionFromTags = (tags: [{ key: string; value: string }]
   if (values && values.length > 0) {
     return values.join(',')
   } else {
-    return MAP_ATTRIBUTION
+    return attribution
   }
 }
