+++ sites/geohub/src/lib/helper/getSelectedTagsFromUrl.ts
@@ -1,11 +1,11 @@
-import { DataCategories, tagSearchKeys } from '$lib/constants'
+import { DataCategories, TagSearchKeys } from '$lib/AppConfig'
 import type { Tag } from '$lib/types'
 import type { Breadcrumb } from '@undp-data/svelte-undp-design'
 import { getBulmaTagColor } from './getBulmaTagColor'
 
 export const getSelectedTagsFromUrl = (url: URL) => {
   const selectedTags: Tag[] = []
-  tagSearchKeys.forEach((key) => {
+  TagSearchKeys.forEach((key) => {
     const values = url.searchParams.getAll(key.key)
     values.forEach((v) => {
       if (selectedTags.find((t) => t.key === key.key && t.value === v)) return
