+++ sites/geohub/src/lib/helper/index.ts
@@ -23,7 +23,6 @@ export * from './sleep'
 export * from './getMaxValueOfCharsInIntervals'
 export * from './distinct'
 export * from './loadMap'
-export * from './createHeaderLinks'
 export * from './initTippy'
 export * from './getBulmaTagColor'
 export * from './createAttributionFromTags'
