+++ sites/geohub/src/lib/server/helpers/getStaticPbfMetadataJson.ts
@@ -3,7 +3,8 @@ import type { VectorLayerTileStatLayer, VectorTileMetadata } from '$lib/types'
 import * as pmtiles from 'pmtiles'
 import arraystat from 'arraystat'
 import { mean, std, median } from 'mathjs'
-import { MAP_ATTRIBUTION, UNIQUE_VALUE_THRESHOLD } from '$lib/constants'
+import { UNIQUE_VALUE_THRESHOLD } from '$lib/constants'
+import { attribution } from '$lib/AppConfig'
 
 /**
  * get metadata json of static pbf
@@ -26,7 +27,7 @@ export const getStaticPbfMetadataJson = async (origin: string, url: string) => {
       bounds: bounds.join(','),
       minzoom: header.minZoom,
       maxzoom: header.maxZoom,
-      attribution: metadata.attribution ?? MAP_ATTRIBUTION,
+      attribution: metadata.attribution ?? attribution,
       description: metadata.description,
       type: metadata.type,
       version: metadata.version,
@@ -79,7 +80,7 @@ export const getStaticPbfMetadataJson = async (origin: string, url: string) => {
     }
 
     if (!data.attribution) {
-      data.attribution = MAP_ATTRIBUTION
+      data.attribution = attribution
     }
 
     const layers = await getStats(origin, pbfpath, data.json.tilestats.layers)
