+++ sites/geohub/src/routes/api/style/[id]/+server.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from './$types'
 import { getStyleById } from '$lib/server/helpers'
-import { AccessLevel } from '$lib/constants'
+import { AccessLevel } from '$lib/AppConfig'
 import DatabaseManager from '$lib/server/DatabaseManager'
 
 export const GET: RequestHandler = async ({ params, locals }) => {
