+++ sites/geohub/src/routes/maps/+page.server.ts
@@ -1,8 +1,9 @@
 import type { PageServerLoad } from './$types'
 import { getMapStats } from '$lib/server/helpers'
-import { AccessLevel, DEFAULT_LIMIT, MapOrderByOptions } from '$lib/constants'
+import { DEFAULT_LIMIT } from '$lib/constants'
 import type { DashboardMapStyle, Pages, StacLink } from '$lib/types'
 import { redirect } from '@sveltejs/kit'
+import { AccessLevel, MapSortingColumns } from '$lib/AppConfig'
 
 export const load: PageServerLoad = async (event) => {
   const { locals, url } = event
@@ -13,7 +14,7 @@ export const load: PageServerLoad = async (event) => {
   // reset default query params if it is not in queryparams
   const sortby = url.searchParams.get('sortby')
   if (!sortby) {
-    apiUrl.searchParams.set('sortby', MapOrderByOptions[0].value)
+    apiUrl.searchParams.set('sortby', MapSortingColumns[0].value)
   }
   const limit = url.searchParams.get('limit')
   if (!limit) {
